==> 2_Feature Extraction.ipynb <==
# Import Libraries

import pandas as pd
import datasist as ds
import numpy as np
from pydub import AudioSegment
import math
import librosa
from sklearn.model_selection import train_test_split
import tensorflow.keras as keras
import matplotlib.pyplot as plt
import librosa.display
from sklearn import preprocessing
from keras.layers import Dense, Dropout
from keras.wrappers.scikit_learn import KerasClassifier
from sklearn.model_selection import GridSearchCV
from keras.callbacks import EarlyStopping
from keras.models import Sequential
from keras.layers import Dense
from keras.optimizers import Adam
from sklearn.metrics import accuracy_score
from sklearn.utils import shuffle

# Read Dataset

df=pd.read_csv('augmented_audio/filtered_data.csv')
df.head()

df.shape

# Extract Spectral Features

SAMPLE_RATE = 22050
num_segments=5
TRACK_DURATION = 20 # measured in seconds

SAMPLES_PER_TRACK = SAMPLE_RATE * TRACK_DURATION
samples_per_segment = int(SAMPLES_PER_TRACK / num_segments)

lists = [[] for _ in range(19)]
folder = 'augmented_audio/recordings/'
filenames=df['filename']

for i,f in enumerate(filenames):
    try:
        print(i)
        # load audio file
        file_path = folder + f + '.wav'
        audio = AudioSegment.from_file(file_path)

        if(audio.duration_seconds>=20):
            y,sr = librosa.load(file_path,offset=0,duration=20)

            # process all segments of audio file
            for d in range(num_segments):

                # calculate start and finish sample for current segment
                start = samples_per_segment * d
                finish = start + samples_per_segment
                y_new=y[start:finish]

            ########################################################
                mfcc = librosa.feature.mfcc(y=y_new, sr=sr, n_mfcc=13)
                for j in range(13):
                    lists[j].append(np.mean(mfcc[j]))
            ########################################################
                chroma_stft = librosa.feature.chroma_stft(y=y_new, sr=sr)
                lists[13].append(np.mean(chroma_stft))
            ########################################################
                spec_cent = librosa.feature.spectral_centroid(y=y_new, sr=sr)
                lists[14].append(np.mean(spec_cent))
            ########################################################
                spec_bw = librosa.feature.spectral_bandwidth(y=y_new, sr=sr)
                lists[15].append(np.mean(spec_bw))
            ########################################################
                rolloff = librosa.feature.spectral_rolloff(y=y_new, sr=sr)
                lists[16].append(np.mean(rolloff))
            ########################################################
                zcr = librosa.feature.zero_crossing_rate(y_new)
                lists[17].append(np.mean(zcr))

                lists[18].append(int(df['label'].loc[df['filename']==f]))
                       
    except:
        print('not done')
        pass

c=['mfcc_'+str(i) for i in range(1,14)]
c=c+['chroma_stft' ,
     'spectral_centroid',
     'spectral_bandwidth',
     'rolloff',
     'zero_crossing_rate',
    'label']

df=pd.DataFrame(lists).T
df.columns=c
df['label']=df['label'].astype(int)
for i in range(5):
    df = shuffle(df)
df.to_csv('Original_Features.csv',index=False)
df.head()

# Feature Scaling

* Feature Scaling is a technique to standardize the independent features present in the data in a fixed range. 
* Feature Scaling may/may not improve the performance of a ML algorithm.
* Normalization and Standardization are two of the most commonly used feature scaling techniques in ML.
* There is no hard and fast rule to tell you when to normalize or standardize your data. One can always start by fitting your model to raw, normalized and standardized data and compare the performance for best results.

## Normalization

scaler = preprocessing.MinMaxScaler()
df2 = scaler.fit_transform(np.array(df.iloc[:, :-1], dtype = float))
df2 = pd.DataFrame(df2,columns=df.columns[:-1])
df2['label']=df['label']
for i in range(5):
    df2 = shuffle(df2)
df2.to_csv('Normalized_Features.csv',index=False)
df2.head()

## Standardization

scaler = preprocessing.StandardScaler()
df2 = scaler.fit_transform(np.array(df.iloc[:, :-1], dtype = float))
df2 = pd.DataFrame(df2,columns=df.columns[:-1])
df2['label']=df['label']
for i in range(5):
    df2 = shuffle(df2)
df2.to_csv('Standardized_Features.csv',index=False)
df2.head()
