==> 6_Model_ML_Standardized.ipynb <==
import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split, cross_val_score
from sklearn.metrics import accuracy_score, confusion_matrix, classification_report 
from pycm import ConfusionMatrix
from sklearn.linear_model import LogisticRegression
from sklearn.tree import DecisionTreeClassifier 
from sklearn.linear_model import Perceptron
from sklearn.ensemble import RandomForestClassifier
from sklearn.neural_network import MLPClassifier
from sklearn.svm import SVC
import pickle
import warnings
warnings.filterwarnings('ignore')

data=pd.read_csv('Standardized_Features.csv')
data.head()

labels = ['Logistic Regression', 'Decision Tree (Gini)','Decision Tree (Entropy)', 'SVM Classifier', 'Random Forest','Perceptron', 'MLP Classifier']
clf1 = LogisticRegression()
clf2=DecisionTreeClassifier(criterion = "gini")
clf3=DecisionTreeClassifier(criterion = "entropy")
clf4=SVC()
clf5=RandomForestClassifier()
clf6=Perceptron()
clf7=MLPClassifier()
models=[clf1,clf2,clf3,clf4,clf5,clf6,clf7]

def train(X,y=data['label']):
    for clf, label in zip(models, labels):
        scores = cross_val_score(clf, X, y, cv=10,scoring='accuracy')
        print("Accuracy: %0.2f (+/- %0.2f) [%s]"
              % (scores.mean(), scores.std(), label))

# On MFCC features only
train(data.iloc[:,:13])

# On All features except MFCC
train(data.iloc[:,13:-1])

# On All features
train(data.iloc[:,:-1])

# From above, we see that SVC, Random Forest and MLP Classifier give best results
# We get best results on All features

# Function to calculate accuracy 
def cal_accuracy(y_test, y_pred): 
    print ("Accuracy : ", accuracy_score(y_test,y_pred)*100) 
    print("Report : ", classification_report(y_test, y_pred)) 
    print(ConfusionMatrix(actual_vector=list(y_test),predict_vector=list(y_pred)))

# MLP
X_train, X_test, y_train, y_test = train_test_split(data.iloc[:,:-1],data['label'], test_size=0.10, random_state=42)
clf7.fit(X_train, y_train)
y_pred = clf7.predict(X_test)
cal_accuracy(y_test, y_pred)

# Random Forest
clf5.fit(X_train, y_train)
y_pred = clf5.predict(X_test)
cal_accuracy(y_test, y_pred)

# SVC
clf4.fit(X_train, y_train)
y_pred = clf4.predict(X_test)
cal_accuracy(y_test, y_pred)

# save the model to disk
filename = 'final_AD_Model.sav'
pickle.dump(clf7, open(filename, 'wb'))
 
# some time later...
 
# load the model from disk
loaded_model = pickle.load(open(filename, 'rb'))
result = loaded_model.score(X_test, y_test)
print(result)

# Observations

* It is known that Feature Scaling improves performance of:
    1. Gradient Descent Based Algorithms (like  linear regression, logistic regression, and neural network, etc.)
    2. Distance-Based Algorithms (like KNN, K-means, and SVM, etc.)
    
* It is also known that Tree-Based Algorithms (like Decision Tree, and Random Forest, etc.) are fairly insensitive/invariant to the scale of the features.
    
* Let's see if these two properties hold true in our case or not!!

## Only MFCCs

|              | Logistic Regression | SVM Classifier | Perceptron | MLP Classifier |
|--------------|---------------------|----------------|------------|----------------|
| Original     | 62                  | 64             | 56         | 69             |
| Normalized   | 55                  | 55             | 51         | 55             |
| Standardized | 62                  | 78             | 53         | 78             |

|              | Decision Tree (Gini) | Decision Tree (Entropy) | Random Forest |
|--------------|----------------------|-------------------------|---------------|
| Original     | 72                   | 72                      | 79            |
| Normalized   | 51                   | 51                      | 52            |
| Standardized | 71                   | 72                      | 79            |

## Other 5 Features

|              | Logistic Regression | SVM Classifier | Perceptron | MLP Classifier |
|--------------|---------------------|----------------|------------|----------------|
| Original     | 55                  | 55             | 52         | 54             |
| Normalized   | 55                  | 55             | 52         | 55             |
| Standardized | 57                  | 59             | 51         | 61             |

|              | Decision Tree (Gini) | Decision Tree (Entropy) | Random Forest |
|--------------|----------------------|-------------------------|---------------|
| Original     | 58                   | 58                      | 61            |
| Normalized   | 50                   | 51                      | 52            |
| Standardized | 57                   | 58                      | 61            |

## All Features

|              | Logistic Regression | SVM Classifier | Perceptron | MLP Classifier |
|--------------|---------------------|----------------|------------|----------------|
| Original     | 61                  | 55             | 53         | 55             |
| Normalized   | 55                  | 55             | 52         | 55             |
| Standardized | 62                  | 78             | 54         | 80             |

|              | Decision Tree (Gini) | Decision Tree (Entropy) | Random Forest |
|--------------|----------------------|-------------------------|---------------|
| Original     | 72                   | 73                      | 79            |
| Normalized   | 51                   | 51                      | 52            |
| Standardized | 72                   | 73                      | 79            |

* It is evident from the tables above that, ML models when trained on:
1. scaled (standardized to be specific) features, perform a little better than when trained on unscaled features (for Gradient Descent Based and Distance-Based algorithms)
2. unscaled features, perform better/equally than when trained on scaled features (for Tree-Based Algorithms)
