==> 7_Model_AutoML.ipynb <==
from tpot import TPOTClassifier
from sklearn.model_selection import train_test_split
import pandas as pd

df=pd.read_csv('Original_Features.csv')
df=df.drop(['rolloff','zero_crossing_rate'],axis=1)
df

# All Features
X=df.iloc[:,:-1]
y=df.iloc[:,-1]
X_train, X_test, y_train, y_test = train_test_split(X,y,test_size=0.10, random_state=42)
tpot = TPOTClassifier(generations=7, population_size=70, verbosity=2, random_state=42, n_jobs=-1)
tpot.fit(X_train, y_train)
#tpot.export('tpot_exported_pipeline.py')
print(tpot.score(X_test, y_test))
tpot.export('tpot_pipeline.py')

# %load "tpot_pipeline.py"
import numpy as np
import pandas as pd
from sklearn.ensemble import ExtraTreesClassifier
from sklearn.model_selection import train_test_split

# NOTE: Make sure that the outcome column is labeled 'target' in the data file
tpot_data = df
features = tpot_data.iloc[:,:-1]
training_features, testing_features, training_target, testing_target = \
            train_test_split(features, df.iloc[:,-1], random_state=42)

# Average CV score on the training set was: 0.8605145778768563
exported_pipeline = ExtraTreesClassifier(bootstrap=False, criterion="gini", max_features=0.7500000000000001, min_samples_leaf=2, min_samples_split=6, n_estimators=100)
# Fix random state in exported estimator
if hasattr(exported_pipeline, 'random_state'):
    setattr(exported_pipeline, 'random_state', 42)

exported_pipeline.fit(training_features, training_target)
y_pred = exported_pipeline.predict(testing_features)

from sklearn.metrics import accuracy_score, confusion_matrix, classification_report 
from pycm import ConfusionMatrix
# Function to calculate accuracy 
def cal_accuracy(y_test, y_pred): 
    print ("Accuracy : ", accuracy_score(y_test,y_pred)*100) 
    print("Report : ", classification_report(y_test, y_pred)) 
    print(ConfusionMatrix(actual_vector=list(y_test),predict_vector=list(y_pred)))

cal_accuracy(testing_target, y_pred)
